refactor componenets
